Completed the post job feature to firestore db
